# assets/Example-Target-Access-Control-swagger.yaml
openapi: 3.0.4
info:
  title: HelloID Access Control target system example API
  description: 'This example API specifies example calls for developing a new API for an Access Control System that will be used for user provisioning from HelloID.<br>Our Starting point, we grant permissions to user accounts and the permissions assigned to a user will be applied to the accessKeys assigned to that user.</br><br>See the Github repo linked below to download the source code.</br>'
  contact:
    name: Tools4everBV Github
    url: https://github.com/orgs/Tools4everBV/
  version: '1.0'
paths:
...
  /api/RoleAuthorizations:
    get:
      tags:
        - RoleAuthorizations
      summary: Get all authorizations for all users
      description: "<h2>Implementation notes</h2>\r\n            We need to retrieve information about all authorizations to support our import entitlement feature, enable reconciliation, and ultimately ensure proper governance."
      operationId: GetAllRoleAuthorizations
      responses:
        '200':
          description: Returns the list of all authorizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleAuthorization'
        '401':
          description: Authentication required or failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    post:
      tags:
        - RoleAuthorizations
      summary: Add authorization
      description: "<h2>Implementation notes</h2>\r\n            We will use this action when an authorization is granted to a user. Since we do not store the result in HelloID, this action does not require a response\r\n            \r\n</br>\r\n            Example:\r\n              \r\n                POST /roleauthorization\r\n                {\r\n                   \"roleId\": 1,\r\n                   \"userId\": 1\r\n                }"
      operationId: AddRoleAuthorization
      requestBody:
        description: The authorization that will be added.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleAuthorization'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleAuthorization'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      tags:
        - RoleAuthorizations
      summary: Delete authorization
      description: "<h2>Implementation notes</h2>\r\n            We will use this action when an authorization is revoked from a user. This action does not require a response. A [204 No Content] is sufficient."
      operationId: DeleteRoleAuthorizationByValue
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
            format: int32
        - name: roleId
          in: query
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: ''
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/api/RoleAuthorizations/{userId}':
    get:
      tags:
        - RoleAuthorizations
      summary: Get all authorizations for a specific user
      description: "<h2>Implementation notes</h2>\r\n            This returns all authorizations currently assigned to the given user.\r\n            \r\n</br>\r\n            Useful when validating which permissions/roles the user has at a given moment."
      operationId: GetRoleAuthorizationsByUserId
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Returns a list of authorizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleAuthorization'
        '404':
          description: If the user has no authorizations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Authentication required or failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/api/RoleAuthorizations/{id}':
    delete:
      tags:
        - RoleAuthorizations
      summary: Delete authorization (by Id)
      description: "<h2>Implementation notes</h2>\r\n            We will use this action when an authorization is revoked from a user. This action does not require a response. A [204 No Content] is sufficient."
      operationId: DeleteAuthorization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: The authorization was successfully removed.
        '401':
          description: Authentication required or failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: No authorization with the specified Id was found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/Roles:
    get:
      tags:
        - Roles
      summary: Get roles
      description: "<h2>Implementation notes</h2>\r\n            Before we can assign a role to a user, we first need to retrieve all available roles. Then, we build out our business rules to, ultimately grant authorizations to users based on information coming from an HR source."
      operationId: GetAllRoles
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    post:
      tags:
        - Roles
      summary: Add Role (optional test)
      description: "<h2>Implementation notes</h2>\r\n            We will use this action only for testing this connector, generally we do not dynamically create roles with helloID \r\n            \r\n</br>\r\n            Example:\r\n              \r\n                POST /roles\r\n                {\r\n                   \"displayName\": \"Admin\"\r\n                   \r\n                }"
      operationId: Add
      requestBody:
        description: The authorization that will be added.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Role'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  schemas:
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
    Role:
      type: object
      properties:
        id:
          type: integer
          description: "This is the internal / database Id.\r\n\r\n\r\nTypically this value will be set by the application itself.\r\n</br>"
          format: int32
          readOnly: true
        displayName:
          type: string
          description: The DisplayName of the role
          nullable: true
          example: Admin
      additionalProperties: false
      description: "The roles that are available in the target system. Like; 'Cleaner', 'Security' or 'Employee'.\r\nA Role of an Access Control System usual contains a set of doors (possible with timeslots) that the user allowed to open.\r\nThis implemntation assumes the roles are granted to the user, and the accessKey are assgined to the user. \r\nThe granted roles are applied to the accessKeys assigned to a user."
    RoleAuthorization:
      required:
        - roleId
        - userId
      type: object
      properties:
        id:
          type: integer
          description: "This is the internal / database Id.\r\n\r\n\r\nTypically this value will be set by the application itself.\r\n</br>"
          format: int32
          readOnly: true
        roleId:
          type: integer
          description: The Id of the role that will be assigned.
          format: int32
          example: 1
        userId:
          type: integer
          description: The Id of the user who will be assigned to this authorization.
          format: int32
          example: 1
      additionalProperties: false
      description: "A user can have multiple authorizations. Each authorization will only have one role Like; 'Cleaner', 'Security' or 'Employee'.\r\n\r\n\r\nA user will be assigned (granted) to an authorization.\r\n</br>"
...
  securitySchemes:
    Bearer:
      type: http
      description: The bearer token
      scheme: Bearer
      bearerFormat: JWT
security:
  - Bearer: [ ]